fix(ci): resolve perf baseline and sonar regressions
